[java-api] Rename JoinOperatorBase to InnerJoinOperatorBase and AbstractJoinOperatorBase to JoinOperatorBase
