+++ flink-core/src/main/java/org/apache/flink/api/common/io/ReplicatingInputFormat.java
@@ -21,6 +21,7 @@ package org.apache.flink.api.common.io;
 
 import org.apache.flink.api.common.functions.RuntimeContext;
 import org.apache.flink.api.common.io.statistics.BaseStatistics;
+import org.apache.flink.api.common.operators.base.InnerJoinOperatorBase;
 import org.apache.flink.configuration.Configuration;
 import org.apache.flink.core.io.InputSplit;
 import org.apache.flink.core.io.InputSplitAssigner;
@@ -33,7 +34,7 @@ import java.io.IOException;
  * This is done by assigning all {@link org.apache.flink.core.io.InputSplit}s generated by the
  * replicated InputFormat to each parallel instance.
  *
- * Replicated data can only be used as input for a {@link org.apache.flink.api.common.operators.base.JoinOperatorBase} or
+ * Replicated data can only be used as input for a {@link InnerJoinOperatorBase} or
  * {@link org.apache.flink.api.common.operators.base.CrossOperatorBase} with the same parallelism as the DataSource.
  * Before being used as an input to a Join or Cross operator, replicated data might be processed in local pipelines by
  * by Map-based operators with the same parallelism as the source. Map-based operators are
@@ -54,7 +55,7 @@ import java.io.IOException;
  *
  * @see org.apache.flink.api.common.io.InputFormat
  * @see org.apache.flink.api.common.io.RichInputFormat
- * @see org.apache.flink.api.common.operators.base.JoinOperatorBase
+ * @see InnerJoinOperatorBase
  * @see org.apache.flink.api.common.operators.base.CrossOperatorBase
  * @see org.apache.flink.api.common.operators.base.MapOperatorBase
  * @see org.apache.flink.api.common.operators.base.FlatMapOperatorBase
