+++ flink-optimizer/src/test/java/org/apache/flink/optimizer/BranchingPlansCompilerTest.java
@@ -25,7 +25,7 @@ import java.util.Set;
 
 import org.apache.flink.api.common.functions.CoGroupFunction;
 import org.apache.flink.api.common.operators.Operator;
-import org.apache.flink.api.common.operators.base.AbstractJoinOperatorBase;
+import org.apache.flink.api.common.operators.base.JoinOperatorBase;
 import org.apache.flink.api.java.io.DiscardingOutputFormat;
 import org.apache.flink.api.java.io.TextOutputFormat;
 import org.apache.flink.api.java.operators.DeltaIteration;
@@ -391,7 +391,7 @@ public class BranchingPlansCompilerTest extends CompilerTestBase {
 			DataSet<Long> map3 = map2.map(new IdentityMapper<Long>()).name("Map 3");
 
 			DataSet<Long> join2 = reduce1.union(reduce2).union(map2).union(map3)
-					.join(map2, AbstractJoinOperatorBase.JoinHint.REPARTITION_SORT_MERGE).where("*").equalTo("*")
+					.join(map2, JoinOperatorBase.JoinHint.REPARTITION_SORT_MERGE).where("*").equalTo("*")
 					.with(new IdentityJoiner<Long>()).name("Join 2");
 
 			join2.output(new DiscardingOutputFormat<Long>());
@@ -889,7 +889,7 @@ public class BranchingPlansCompilerTest extends CompilerTestBase {
 						.withBroadcastSet(bc_input2, "bc2");
 			
 			DataSet<Tuple2<Long, Long>> joinResult = joinInput1
-				.join(joinInput2, AbstractJoinOperatorBase.JoinHint.REPARTITION_HASH_FIRST).where(0).equalTo(1)
+				.join(joinInput2, JoinOperatorBase.JoinHint.REPARTITION_HASH_FIRST).where(0).equalTo(1)
 				.with(new DummyFlatJoinFunction<Tuple2<Long,Long>>());
 			
 			input
