+++ flink-staging/flink-ml/src/main/scala/org/apache/flink/ml/recommendation/ALS.scala
@@ -19,7 +19,7 @@
 package org.apache.flink.ml.recommendation
 
 import java.{util, lang}
-import org.apache.flink.api.common.operators.base.AbstractJoinOperatorBase.JoinHint
+import org.apache.flink.api.common.operators.base.JoinOperatorBase.JoinHint
 import org.apache.flink.api.scala._
 import org.apache.flink.api.common.operators.Order
 import org.apache.flink.core.memory.{DataOutputView, DataInputView}
